refactor(cli-utils): move unified diff logic to @markuplint/cli-utils
Move the LCS-based unified diff algorithm from markuplint CLI into
@markuplint/cli-utils as a reusable `unifiedDiff()` function.
The CLI dry-run output now delegates to this shared utility.
